* remove enums package. * complete isNullOrEmpty method.
