+++ angular-beans/src/main/java/angularBeans/api/Eval.java
@@ -5,7 +5,7 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-import angularBeans.enums.Callback;
+import angularBeans.events.Callback;
 
 @Target(ElementType.METHOD)
 @Retention(RetentionPolicy.RUNTIME)
