+++ angular-beans/src/main/java/angularBeans/boot/Callback.java
@@ -1,0 +1,20 @@
+package angularBeans.boot;
+
+/**
+ * Class enum of Callbaks.
+ * 
+ * @author Bessem Hmidi, melbehi
+ *
+ */
+public enum Callback {
+
+	/**
+	 * BEFORE_SESSION_READY.
+	 */
+	BEFORE_SESSION_READY,
+	
+	/**
+	 * AFTER_SESSION_READY.
+	 */
+	AFTER_SESSION_READY
+}
